pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-multiverse"
version = "0.0.3"
authors = [
  { name="Emma Laurijssens van Engelenhoven", email="emma@emma.ee" },
]
description = "Pixel pusher for Pimoroni Multiverse LED products"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial",
    "pillow"
]

[project.urls]
Homepage = "https://github.com/elaurijssens/pixel-multiverse"
Issues = "https://github.com/elaurijssens/pixel-multiverse/issues"